pyproject.toml
[project]
name = "subtitles-ocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytesseract>=0.3.13",
    "pillow>=11.2.1",
    "ffmpeg-python>=0.2.0",
    "rich>=14.0.0",
    "numpy",
]
